requirements.txt
instagramy==4.5
igramscraper==0.3.5
python-dotenv==0.20.0